+++ tscrop/tscrop.go
@@ -2,13 +2,13 @@ package main
 
 import (
 	"bufio"
+	"bytes"
 	"errors"
 	"flag"
 	"fmt"
 	"github.com/anthonynsimon/bild/imgio"
 	"github.com/borevitzlab/go-timestreamtools/utils"
 	"github.com/oliamb/cutter"
-	"github.com/rwcarlsen/goexif/exif"
 	"golang.org/x/image/tiff"
 	"image"
 	"io"
@@ -23,12 +23,11 @@ import (
 )
 
 var (
-	errLog                              *log.Logger
-	rootDir, outputDir, targetExtension string
-	corner1, corner2, gridxy, chunkSize image.Point
-	imageEncoder                        imgio.Encoder
-	center                      		bool
-
+	errLog                                             *log.Logger
+	rootDir, outputDir, outfmt, infmt, targetExtension string
+	corner1, corner2, gridxy, chunkSize                image.Point
+	imageEncoder                                       imgio.Encoder
+	center                                             bool
 )
 
 // TIFFEncoder returns an encoder to the Tagged Image Format
@@ -49,45 +48,33 @@ func minMax(a, b int) (min, max int) {
 	return
 }
 
-func getExifData(thisFile string) ([]byte, error) {
-	fileHandler, err := os.Open(thisFile)
-	if err != nil {
-		// file wouldnt open
-		return []byte{}, err
-	}
+func cropImage(sourceImg utils.Image, destPath string) (err error) {
 
-	exifData, err := exif.Decode(fileHandler)
-	if err != nil {
-		// exif wouldnt decode
-		return []byte{}, err
-	}
-
-	jsonBytes, err := exifData.MarshalJSON()
-	if err != nil {
-		return []byte{}, err
-	}
-	return jsonBytes, nil
-}
-
-func writeExifJSON(sourcePath, destPath string) {
-	exifJSON, exifJSONErr := getExifData(sourcePath)
-
-	if exifJSONErr != nil {
-		errLog.Printf("[exif] couldnt read data from %s", sourcePath)
-	}
-	if len(exifJSON) > 0 {
-		exifJSONErr := ioutil.WriteFile(destPath+".json", exifJSON, 0644)
-		if exifJSONErr != nil {
-			errLog.Printf("[exif] couldnt write json %s", destPath)
+	if len(sourceImg.Data) == 0 {
+		file, err := os.Open(sourceImg.Path)
+		if err != nil {
+			return
 		}
-	}
-}
+		// read the image bytes into the img.Data
+		buf1 := new(bytes.Buffer)
+		_, err = buf1.ReadFrom(file)
+		if err != nil {
+			return
+		}
 
-func cropImage(sourcePath, destPath string) (err error) {
-	img, err := imgio.Open(sourcePath)
+		sourceImg.Data = buf1.Bytes()
+		buf1.Reset()
+	}
+
+	imgReader := bytes.NewReader(sourceImg.Data)
+	img, _, err := image.Decode(imgReader)
 	if err != nil {
 		return
 	}
+	if img == nil {
+		return errors.New("[imgload] nil img wtf")
+	}
+
 	var cropImage image.Image
 	if center {
 		cropImage, err = cutter.Crop(img, cutter.Config{
@@ -104,7 +91,7 @@ func cropImage(sourcePath, destPath string) (err error) {
 		})
 	}
 	var wg sync.WaitGroup
-	wg.Add(gridxy.X * gridxy.Y)
+	wg.Add(gridxy.X * gridxy.Y) // add this number to the waitgroup so wait for all of these to finish.
 
 	for xPos := 0; xPos < gridxy.X; xPos++ {
 		for yPos := 0; yPos < gridxy.Y; yPos++ {
@@ -121,18 +108,37 @@ func cropImage(sourcePath, destPath string) (err error) {
 					errLog.Printf("[crop] error cropping: %s", cropErr)
 					return
 				}
+				buf2 := new(bytes.Buffer)
+				imgWriter := bufio.NewWriter(buf2)
+				err = imageEncoder(imgWriter, cropped)
+				if err != nil {
+					return
+				}
+
+				imgWriter.Flush()
+
+				// read the image bytes into the img.Data
 
 				destPos := fmt.Sprintf("%d,%d", xPos, yPos)
 				destPath := fmt.Sprintf(destPath, destPos)
-				os.MkdirAll(path.Dir(destPath), 0755)
+				cImg := utils.Image{
+					OriginalPath:  sourceImg.OriginalPath,
+					Data:          buf2.Bytes(),
+					Timestamp:     sourceImg.Timestamp,
+					ExifTimestamp: sourceImg.ExifTimestamp,
+					CmdList:       append(sourceImg.CmdList, strings.Join(os.Args, " ")),
+				}
+				buf2.Reset()
 
-				writeErr := imgio.Save(destPath, cropped, imageEncoder)
-				if writeErr != nil {
+				// write image out.
+				if writeErr := utils.WriteImageToFile(cImg, destPath); writeErr != nil {
 					errLog.Printf("[crop] error saving crop: %s", cropErr)
 					return
 				}
+
 				// output the relative image path
-				utils.EmitPath(destPath)
+				utils.Emit(cImg, outfmt)
+
 			}(xPos, yPos)
 		}
 	}
@@ -141,22 +147,25 @@ func cropImage(sourcePath, destPath string) (err error) {
 	return
 }
 
-func visit(filePath string, info os.FileInfo, _ error) error {
-	// skip directories
+
+
+func visitWalk(filePath string, info os.FileInfo, _ error) error {
 	// skip directories
 	if info.IsDir() {
 		return nil
 	}
-
-	if path.Ext(filePath) == ".json" {
-		return nil
+	img, err := utils.LoadImage(filePath)
+	img.OriginalPath = filePath
+	if err != nil {
+		errLog.Printf("[load] %s", err)
 	}
 
-	if strings.HasPrefix(filepath.Base(filePath), "."){
-		return nil
-	}
+	return visit(img)
+}
 
-	ext := path.Ext(filePath)
+func visit(img utils.Image) error {
+
+	ext := path.Ext(img.Path)
 	switch extlower := strings.ToLower(ext); extlower {
 	case ".jpeg", ".jpg", ".tif", ".tiff", ".cr2":
 		break
@@ -164,15 +173,14 @@ func visit(filePath string, info os.FileInfo, _ error) error {
 		return nil
 	}
 
-	basePath := path.Base(filePath)
+	basePath := path.Base(img.Path)
 	// parse the new filepath
 	noExtension := strings.TrimSuffix(basePath, ext)
 	newBase := fmt.Sprintf("%s.%s", noExtension, targetExtension)
-	writeExifJSON(filePath, path.Join(outputDir, newBase))
 	newPath := path.Join(outputDir, "%s", newBase)
 
 	// convert the image
-	if err := cropImage(filePath, newPath); err != nil {
+	if err := cropImage(img, newPath); err != nil {
 		errLog.Printf("[crop] %s", err)
 		return nil
 	}
@@ -180,32 +188,33 @@ func visit(filePath string, info os.FileInfo, _ error) error {
 }
 
 var usage = func() {
-	fmt.Printf("usage of %s:\n", os.Args[0])
-	fmt.Println()
-	fmt.Println("centered crop to 1920x1080")
-	fmt.Printf("\t%s -center -c1 1920,1080\n", os.Args[0])
-	fmt.Println("cut out 120,10 to 400,60")
-	fmt.Printf("\t%s -c1 120,10 c2 400,60\n", os.Args[0])
-	fmt.Println("centered crop to 1920x1080 and output to <destination>")
-	fmt.Printf("\t%s -center -c1 1920,1080 -output <destination>\n", os.Args[0])
-	fmt.Println()
-	fmt.Println("flags:")
-	fmt.Println()
-	fmt.Println("\t-center: center the crop, specify width,height with c1")
-	fmt.Println("\t-c1: corner 1 (in pixels, comma separated)")
-	fmt.Println("\t-c2: corner 2 (in pixels, comma separated, ignored if center is specified)")
-	fmt.Println("\t-grid: split the area into this many equal crops (default=1,1)")
-	fmt.Println("\t-type: set the output image type (default=jpeg)")
-	fmt.Println("\t\tavailable image types:")
-	fmt.Println()
-	fmt.Println("\t\tjpeg, png")
-	fmt.Println("\t\ttiff: tiff with Deflate compression (alias for tiff-deflate)")
-	fmt.Println("\t\ttiff-none: tiff with no compression")
-	fmt.Println("\t-output: set the <destination> directory (default=<cwd>/<crop>)")
-	fmt.Println()
-	fmt.Println("reads filepaths from stdin")
-	fmt.Println("writes paths to resulting files to stdout")
-	fmt.Println("will ignore any line from stdin that isnt a filepath (and only a filepath)")
+	use := `
+usage of %s:
+
+flags:
+	-center: center the crop, specify width,height with c1
+	-c1: corner 1 (in pixels, comma separated)
+	-c2: corner 2 (in pixels, comma separated, ignored if center is specified)
+	-grid: split the area into this many equal crops (default=1,1)
+	-type: set the output image type (default=jpeg)
+	-output: set the <destination> directory (default=<cwd>/<crop>)
+
+available image types:
+	jpeg, png
+	tiff: tiff with Deflate compression (alias for tiff-deflate)
+	tiff-none: tiff with no compression
+
+examples:
+	centered crop to 1920x1080:
+		%s -center -c1 1920,1080
+	cut out 120,10 to 400,60:
+		%s -c1 120,10 c2 400,60
+	centered crop to 1920x1080 and output to <destination>:
+		%s -center -c1 1920,1080 -output <destination>
+	4x4 grid crop of centered 1920x1080:
+		%s -grid 4,4  -center -c1 1920,1080 -output <destination>
+`
+	fmt.Printf(use, os.Args[0], os.Args[0], os.Args[0], os.Args[0], os.Args[0])
 }
 
 func stringToPoint(str, sep string) (image.Point, error) {
@@ -231,6 +240,8 @@ func init() {
 	// set flags for flag
 	flag.StringVar(&rootDir, "source", "", "source directory")
 	flag.StringVar(&outputDir, "output", "", "output directory")
+	flag.StringVar(&outfmt, "outfmt", "path", "output format")
+	flag.StringVar(&infmt, "infmt", "path", "input format")
 	flag.BoolVar(&center, "center", false, "center crop")
 	outputType := flag.String("type", "jpeg", "output image type")
 	c1 := flag.String("c1", "0,0", "corner 1")
@@ -300,64 +311,90 @@ func init() {
 	if rootDir != "" {
 		if _, err := os.Stat(rootDir); err != nil {
 			if os.IsNotExist(err) {
-				errLog.Printf("[path] <source> %s does not exist.", rootDir)
+				errLog.Printf("[path] <source> %s does not exist", rootDir)
 				os.Exit(1)
 			}
 		}
-	}
-
-
-	if _, err := os.Stat(outputDir); err != nil {
-		os.MkdirAll(outputDir, 0755)
 	}
 }
 
 func main() {
 
-	if outputDir == "" {
-		tmpDir, err := ioutil.TempDir("", "tscrop-")
+	if outputDir == "tmp" {
+		tmpDir, err := ioutil.TempDir("", "tsresize-")
 		if err != nil {
 			panic(err)
 		}
-		// pass delete dir onto next step once finished
-		defer utils.EmitPath("#-" + tmpDir)
+
+		defer utils.EmitCleanup(tmpDir, outfmt)
+
 		outputDir = tmpDir
 	}
 
-	cropFmtCenter := fmt.Sprintf("%d,%d", corner1.X, corner1.Y)
-	cropFmt := fmt.Sprintf("%d,%d-%d,%d", corner1.X, corner1.Y, corner2.X, corner2.Y)
-
 	if center {
+		cropFmtCenter := fmt.Sprintf("%d,%d", corner1.X, corner1.Y)
 		outputDir = path.Join(outputDir, cropFmtCenter)
 	} else {
+		cropFmt := fmt.Sprintf("%d,%d-%d,%d", corner1.X, corner1.Y, corner2.X, corner2.Y)
 		outputDir = path.Join(outputDir, cropFmt)
 	}
 
+	os.MkdirAll(outputDir, 0755)
 	if rootDir != "" {
-		if err := filepath.Walk(rootDir, visit); err != nil {
+		if err := filepath.Walk(rootDir, visitWalk); err != nil {
 			errLog.Printf("[walk] %s", err)
 		}
 	} else {
-		// start scanner and wait for stdin
-		scanner := bufio.NewScanner(os.Stdin)
-		for scanner.Scan() {
-
-			text := strings.Replace(scanner.Text(), "\n", "", -1)
-			if strings.HasPrefix(text, "[") {
-				errLog.Printf("[stdin] %s", text)
-				continue
-			} else if strings.HasPrefix(text, "#-") {
-				// was signalled deletion of previous tmpdir, wait until finished
-				defer os.RemoveAll(strings.TrimPrefix(text, "#-"))
-			} else {
-				finfo, err := os.Stat(text)
-				if err != nil {
-					errLog.Printf("[stat] %s", text)
+		if infmt == "path" {
+			// start scanner and wait for stdin
+			scanner := bufio.NewScanner(os.Stdin)
+			for scanner.Scan() {
+				text := strings.Replace(scanner.Text(), "\n", "", -1)
+				if strings.HasPrefix(text, "[") {
+					errLog.Printf("[stdin] %s", text)
 					continue
+				} else if strings.HasPrefix(text, "#-") {
+					// was signalled deletion of previous tmpdir, wait until finished
+					defer os.RemoveAll(strings.TrimPrefix(text, "#-"))
+				} else {
+					img, err := utils.LoadImage(text)
+					if err != nil {
+						errLog.Printf("[load] %s", err)
+					}
+					visit(img)
 				}
-				visit(text, finfo, nil)
+				data := strings.Replace(scanner.Text(), "\n", "", -1)
+				if strings.HasPrefix(data, "[") {
+					errLog.Printf("[stdin] %s", data)
+					continue
+				} else {
+					img, err := utils.LoadImage(data)
+					if err != nil {
+						errLog.Printf("[load] %s", err)
+					}
+					visit(img)
+				}
 			}
+
+		} else {
+			//data := scanner.Bytes()
+			//img := utils.Image{}
+			//err := json.Unmarshal(data, &img)
+			//if err != nil {
+			//
+			//	errLog.Printf("[json] %s", err)
+			//	continue
+			//}
+
+			// clean up...
+			//t := utils.TempDir{}
+			//if err := json.Unmarshal(data, &t); err == nil{
+			//	defer fmt.Printf("Removing %s\n", t.Path)
+			//	defer os.RemoveAll(t.Path)
+			//}
+			//continue
+
+			utils.Handle(visit, os.RemoveAll, infmt)
 		}
 	}
-
 }
